authenticted user should not send profile id when working with his profile

+++ src/user/profile/profile.service.ts
@@ -1,9 +1,10 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { Identifier } from 'src/core/abstract-data-layer/types';
 import { IProfile } from 'src/core/entities/profile-entity.abstract';
 import { FileService } from 'src/file/file.service';
 import { AbstractFileStorageService } from 'src/file-storage/file-storage.service.abstract';
 import { FILE_STOREAGE_SERVICE_PROVIDER_TOKEN } from 'src/file-storage/fs-file-storeage.constants';
+import { ERROR_MESSAGES } from 'src/utils/error-messages';
 
 import { UpdateProfileDataDTO } from './dtos/update-profile-data.dto';
 import { PROFILE_REPOSITORY_PROVIDER_TOKEN } from './profile.constants';
@@ -19,22 +20,30 @@ export class ProfileService {
     private readonly fileStorageService: AbstractFileStorageService,
   ) {}
 
+  async getUserProfile(userId: Identifier) {
+    return this.profileRepository.getUserProfile(userId);
+  }
+
   async updateProfileData(
-    profileId: Identifier,
+    userId: Identifier,
     updateProfileDataDTO: UpdateProfileDataDTO,
   ) {
+    const userProfile = await this.getUserProfile(userId);
+
     const { name } = updateProfileDataDTO;
 
-    const updated = await this.profileRepository.updateOneById(profileId, {
+    const updated = await this.profileRepository.updateOneById(userProfile.id, {
       name,
     });
 
     return !!updated;
   }
 
-  async updateProfilePhoto(profileId: number, photoStorageIdentifier: string) {
+  async updateProfilePhoto(userId: Identifier, photoStorageIdentifier: string) {
     try {
-      const profile = await this.profileRepository.getOneById(profileId);
+      const userProfile = await this.getUserProfile(userId);
+
+      const profile = await this.profileRepository.getOneById(userProfile.id);
       profile.visualResource.images.forEach(async (image) => {
         await this.removeProfilePhoto(image.id);
       });
@@ -46,7 +55,16 @@ export class ProfileService {
     }
   }
 
-  async removeProfilePhoto(photoId: Identifier) {
+  async removeProfilePhoto(userId: Identifier) {
+    const userProfile = await this.getUserProfile(userId);
+    const { images } = userProfile.visualResource;
+
+    if (!images.length) {
+      throw new NotFoundException(ERROR_MESSAGES.USER_HAS_NO_PHOTO);
+    }
+
+    const photoId = images[0].id;
+
     const deleted = await this.fileService.deleteOne(photoId);
     return deleted;
   }
